backend/pyproject.toml
[project]
name = "dedupe-it"
version = "0.1.0"
description = "Deduplicate structured data with zero configuration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.37.1",
    "openai>=1.52.2",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.0",
    "sentence-transformers>=3.2.1",
    "httpx>=0.24.1",
    "pyarrow>=17.0.0",
    "numpy>=1.26.4",
    "sqlglot>=25.20.2",
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
    "polars>=1.12.0",
    "duckdb>=1.1.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py312"
